Remove internals of struct Pixa/Boxa/Pta/FPixa/DPix from public API
