+++ prog/hashtest.c
@@ -47,8 +47,9 @@
 #include <config_auto.h>
 #endif  /* HAVE_CONFIG_H */
 
-#include "array_internal.h"
 #include <string.h>
+#include "array_internal.h"
+#include "pix_internal.h"
 
 static SARRAY *BuildShortStrings(l_int32 nchars);
 static PTA *BuildPointSet(l_int32  w, l_int32 h);
