+++ prog/pixaatest.c
@@ -34,7 +34,7 @@
 #include <config_auto.h>
 #endif  /* HAVE_CONFIG_H */
 
-#include "allheaders.h"
+#include "pix_internal.h"
 
 static const l_int32  nx = 10;
 static const l_int32  ny = 12;
