+++ prog/sorttest.c
@@ -35,7 +35,7 @@
 #include <config_auto.h>
 #endif  /* HAVE_CONFIG_H */
 
-#include "allheaders.h"
+#include "pix_internal.h"
 
 int main(int    argc,
          char **argv)
