+++ src/boxfunc4.c
@@ -69,7 +69,7 @@
 #endif  /* HAVE_CONFIG_H */
 
 #include <math.h>
-#include "allheaders.h"
+#include "pix_internal.h"
 
 /*---------------------------------------------------------------------*
  *                     Boxa and boxaa range selection                  *
